Add check for nulls + check for deletionProtection

+++ utils/cleanupResources.ts
@@ -1,4 +1,7 @@
-var dotenv = require('dotenv');
+import dotenv from 'dotenv';
+
+import pinecone from '../dist';
+
 dotenv.config();
 
 for (const envVar of ['PINECONE_API_KEY']) {
@@ -8,23 +11,27 @@ for (const envVar of ['PINECONE_API_KEY']) {
     console.log(`INFO Found environment variable ${envVar} in .env file`);
   }
 }
-
-var pinecone = require('../dist');
-
 (async () => {
   const p = new pinecone.Pinecone();
 
   const collectionList = await p.listCollections();
-  for (const collection of collectionList.collections) {
-    console.log(`Deleting collection ${collection.name}`);
-    await p.deleteCollection(collection.name);
+  if (collectionList.collections) {
+    for (const collection of collectionList.collections) {
+      console.log(`Deleting collection ${collection.name}`);
+      await p.deleteCollection(collection.name);
+    }
   }
 
   const response = await p.listIndexes();
-  for (const index of response.indexes) {
-    console.log(`Deleting index ${index.name}`);
-    await p.deleteIndex(index.name);
-  }
+  if (response.indexes) {
+    for (const index of response.indexes) {
+      if (index.deletionProtection == 'enabled') {
+        index.deletionProtection = 'disabled';
+      }
+      console.log(`Deleting index ${index.name}`);
+      await p.deleteIndex(index.name);
+    }
 
-  process.exit();
+    process.exit();
+  }
 })();
